PerfMonTests: Fix cppcheck warnings.
signed/unsigned conversion.

+++ Control/PerformanceMonitoring/PerfMonTests/src/PerfMonTestMallocAlg.cxx
@@ -94,7 +94,7 @@ StatusCode MallocAlg::execute()
       if ( evtnbr >= 2*m_evtNbr ) {
         std::vector<ElemType> c2_array(nmax);
         for ( std::size_t i = 0; i!=nmax; ++i ) {
-          c2_array[i] =  -1*i;
+          c2_array[i] =  -1*static_cast<int>(i);
         }
       }
       
@@ -108,7 +108,7 @@ StatusCode MallocAlg::execute()
       if ( evtnbr >= 2*m_evtNbr ) {
         ElemType c2_array[nmax];
         for ( std::size_t i = 0; i!=nmax; ++i ) {
-          c2_array[i] =  -1*i;
+          c2_array[i] =  -1*static_cast<int>(i);
         }
         // dummy stuff to silence gcc-warning.
         if (c2_array[0] > c1_array[0]) {
